NOR-183 splitting out promoted links into own component
extracted ServicesLinkList to own component
