+++ src/library/structure/HomeHero/HomeHero.types.ts
@@ -1,3 +1,4 @@
+import {PromotedLinkProp} from './../../components/PromotedLinks/PromotedLinks.types'
 export interface HomeHeroProps {
    /**
    * The line of text before the logo
@@ -33,13 +34,3 @@ export interface HeroImageProp {
     imageAltText?: string;
 }
 
-export interface PromotedLinkProp {
-    /**
-     * The title of the link
-     */
-    title: string;
-    /**
-     * The url of the link
-     */
-    url: string;
-}
