+++ src/library/structure/ServicesLinksList/ServicesLinksList.stories.tsx
@@ -564,3 +564,9 @@ export const ServicesLinksListWithoutIcons = Template.bind({});
 ServicesLinksListWithoutIcons.args = {
   serviceLinksArray: exampleDataNoIcons
 };
+
+export const ExampleServicesLinksListWithCard = Template.bind({});    
+ExampleServicesLinksListWithCard.args = {
+  serviceLinksArray: exampleData,
+  hasBackground: true
+};
